requirements.txt
Flask==3.0.0
pandas==2.1.1
numpy==1.26.0
openpyxl==3.1.2
SQLAlchemy==2.0.21
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
xlrd==2.0.1
